examples/data-science-analystics/cpu-orchestration/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.22.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
apache-airflow==2.10.4
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.1
apache-airflow-providers-sqlite==3.9.1
apispec==6.8.0
apprise==1.9.6
argcomplete==3.5.2
asgi-lifespan==2.1.0
asgiref==3.8.1
asyncpg==0.31.0
attrs==24.2.0
babel==2.16.0
blinker==1.9.0
cachelib==0.9.0
cachetools==6.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
coolname==2.2.0
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==42.0.8
dateparser==1.2.2
Deprecated==1.2.15
dill==0.3.1.1
dnspython==2.7.0
docker==7.1.0
durationpy==0.10
email_validator==2.2.0
exceptiongroup==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.47.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
graphviz==0.21
greenlet==3.1.1
griffe==1.15.0
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.0
humanize==4.15.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jmespath==0.10.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==31.0.0
lazy-object-proxy==1.10.0
limits==3.14.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.8
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
narwhals==2.15.0
numpy==2.4.1
oauthlib==3.3.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pathspec==0.12.1
pendulum==2.1.2
plotly==6.5.2
pluggy==1.5.0
prefect==2.20.25
prison==0.2.1
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.21
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
pytzdata==2020.1
PyYAML==6.0.2
readchar==4.2.1
referencing==0.35.1
regex==2026.1.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.17.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.16.0
toml==0.10.2
typer==0.17.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
uc-micro-py==1.0.3
ujson==5.11.0
universal_pathlib==0.2.5
urllib3==2.2.3
uvicorn==0.40.0
websocket-client==1.9.0
websockets==13.1
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.17.0
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
